Move Locations to Loc.hs
