+++ app/LSP/MegaparsecToLSP.hs
@@ -5,7 +5,7 @@ import Data.Map (Map)
 import Data.Map qualified as M
 import Text.Megaparsec
     ( SourcePos(SourcePos), unPos )
-import Utils ( Loc(..) )
+import Loc ( Loc(..) )
 import Language.LSP.Types
     ( Position(..),
       Range(..)
