+++ src/Lookup.hs
@@ -31,7 +31,7 @@ import Syntax.RST.Types (PolarityRep(..), Polarity(..))
 import Syntax.CST.Types (PrdCnsRep(..))
 import Syntax.CST.Names
 import Syntax.CST.Kinds (MonoKind)
-import Utils
+import Loc ( Loc, defaultLoc )
 
 ---------------------------------------------------------------------------------
 -- We define functions which work for every Monad which implements:
