+++ src/Resolution/Program.hs
@@ -20,7 +20,7 @@ import Syntax.RST.Types qualified as RST
 import Syntax.RST.Types (Polarity(..), PolarityRep(..))
 import Syntax.CST.Kinds
 import Syntax.CST.Names
-import Utils (Loc, defaultLoc)
+import Loc (Loc, defaultLoc)
 
 
 ---------------------------------------------------------------------------------
