+++ src/Syntax/Core/Terms.hs
@@ -20,7 +20,7 @@ import Data.List (elemIndex)
 import Data.Maybe (fromJust, isJust)
 import Data.Text qualified as T
 import Syntax.Core.Annot
-import Utils
+import Loc
 import Errors
 import Syntax.CST.Types (PrdCns(..), PrdCnsRep(..))
 import Syntax.TST.Terms (ShiftDirection(..))
