+++ src/Syntax/RST/Types.hs
@@ -8,7 +8,7 @@ import Syntax.CST.Kinds ( Variance(..) )
 import Syntax.CST.Types ( PrdCnsRep(..), PrdCns(..), Arity)
 import Syntax.CST.Names
     ( MethodName, RecTVar, RnTypeName, SkolemTVar, UniTVar, XtorName )
-import Utils
+import Loc ( Loc, defaultLoc )
 
 ------------------------------------------------------------------------------
 -- Polarity
