+++ src/Syntax/TST/Types.hs
@@ -10,7 +10,7 @@ import Syntax.CST.Kinds (Variance(..), MonoKind)
 import Syntax.CST.Types ( PrdCnsRep(..), PrdCns(..), Arity)
 import Syntax.CST.Names ( MethodName, RecTVar, RnTypeName, SkolemTVar, UniTVar, XtorName )
 
-import Utils
+import Loc
 
 ------------------------------------------------------------------------------
 -- CovContraList
